+++ packages/common/src/requirements-txt.ts
@@ -1,7 +1,10 @@
+const pipCommentRegexp = /\s#.*$/;
+
 export function parseRequirementsTxt(content: string): string[] {
   return content
     .split("\n")
-    .filter((r) => !r.startsWith("#"))
+    .filter((r) => !r.startsWith("#")) // https://pip.pypa.io/en/latest/reference/requirements-file-format/#comments
+    .map((r) => r.replace(pipCommentRegexp, "")) // https://pip.pypa.io/en/latest/reference/requirements-file-format/#comments
     .map((r) => r.trim())
     .filter((r) => r !== "");
 }