+++ packages/desktop/scripts/build_bin.js
@@ -9,7 +9,6 @@ require("esbuild")
     minify: true,
     platform: "node",
     external: [
-      "@stlite/kernel", // Don't touch `require.resolve("@stlite/kernel") at the bundle time.
       "pyodide", // The `pyodide` package must be installed at runtime for the included Wasm files, so there is no reason to bundle it here.
       "node-fetch", // `node-fetch` will be installed at runtime anyway because it is one dependency of the `pyodide` package, so there is no reason to bundle it here.
       "fs-extra", // `fs-extra` and `yargs` will be installed at runtime anyway as the dependencies of `electron-builder`, so we don't have to bundle them here.